% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{asymmRerootingMethod}
\alias{asymmRerootingMethod}
\title{Calculates ancestral likelihoods of ancestral states for discrete characters}
\usage{
asymmRerootingMethod(tree, tips, Q)
}
\arguments{
\item{tree}{An object of class "phylo"}

\item{tips}{The phenotype data for the species at the tips of the tree. Must be in the same order as tree$tip.label}

\item{Q}{A transition rate matrix fit on the tree and tips}
}
\value{
A matrix of ancestral likelihoods with rows in the same order as the internal nodes in the tree
}
\description{
Calculates ancestral likelihoods of ancestral states for discrete characters
}
