% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr_diagnostics.R
\name{boxPlotTest}
\alias{boxPlotTest}
\title{Compares prediction accuracy under different rate models and represents the results as a series of boxplots}
\usage{
boxPlotTest(
  treesObj,
  phenvals,
  rate_models,
  nsims = 100,
  confidence_threshold = NULL,
  ...
)
}
\arguments{
\item{treesObj}{the trees object returned by readTrees}

\item{phenvals}{the named phenotype vector with names matching the tip labels in the trees object}

\item{rate_models}{a list of rate models to compare}

\item{nsims}{the number of simulations to perform for each model of evolution (each rate model)}

\item{confidence_threshold}{if not NULL, only compares prediction accuracy for nodes whose maximum likelihood is greater than the confidence threshold. Must be between 0 and 1.}
}
\value{
a list of boxplots for each rate model
}
\description{
Compares prediction accuracy under different rate models and represents the results as a series of boxplots
}
