% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{calculateCorPermuted}
\alias{calculateCorPermuted}
\title{Calculate permulation correlation statistics}
\usage{
calculateCorPermuted(
  permulated.paths,
  RERmat,
  min.sp = 10,
  min.pos = 2,
  method = "k"
)
}
\arguments{
\item{permulated.paths}{A nested list of permulated paths (e.g., output of \code{\link{calculatePermulatedPaths_apply}}}

\item{RERmat}{An RER matrix calculated using \code{\link{getAllResiduals}}.}

\item{min.sp}{Minimum number of species that must be present for a gene}

\item{min.pos}{Minimum number of species that must be present in the foreground (non-zero phenotype values)}

\item{method}{Method used to compute correlations. Accepts the same arguments as \code{\link{cor}}. Set to "auto" to select automatically based on the number of unique values in charP. This will also auto set the winsorization for Pearson correlation. Set winsorizetrait=some number and winsorizeRER=some number to override}
}
\value{
A nested list containing the correlation statistics for the permulations
}
\description{
Calculate permulation correlation statistics
}
