% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{categoricalPermulations}
\alias{categoricalPermulations}
\title{Runs permulations for categorical data}
\usage{
categoricalPermulations(
  treesObj,
  phenvals,
  rm,
  rp = "auto",
  ntrees,
  percent_relax = 0
)
}
\arguments{
\item{treesObj}{trees object returned by readTrees}

\item{phenvals}{the named phenotype vector}

\item{rm}{the rate model, it should be the same as the one used to reconstruct the ancestral history of the trait}

\item{rp}{root prior, the default is auto}

\item{ntrees}{the number of null trees to generate}

\item{root_prob}{The probabilities of the different states at the root for the simulations. Can be "flat", "stationary", or a numeric vector of length Nstates. See the root_probabilities parameter under simulate_mk_model in the castor package for more information.}
}
\value{
a set of permulated phenotype trees
}
\description{
Runs permulations for categorical data
}
