% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{char2Paths}
\alias{char2Paths}
\title{turns a named vector of characters into a paths vector to be used with \code{\link{getAllCor}}}
\usage{
char2Paths(tip.vals, treesObj, altMasterTree = NULL, metric = "diff",
  se.filter = -1)
}
\arguments{
\item{tip.vals}{the trait/phenotype/character value at the tip, \code{names(tip.vals)} should match some of the \code{mastertree$tip.label}, though a perfect match is not required}

\item{treesObj}{A treesObj created by \code{\link{readTrees}}}

\item{metric}{The metric used to translate node values into branch values. "Diff" takes the difference and makes the result phylogenetically independent. Other possible values are "mean" (the mean of the two values) and "last" the value of the most recent species on the branch. The last two options are not phylogenetically independent and downstream computations for those are not yet implemented}

\item{se.filter}{Will remove branch values that are not at least \code{se.filter*edge.se} away from 0 (where edge.se is the standard error in the estimate for the edge value). Only implemented for \code{metric="diff"}. Set \code{se.filter} to a positive value to filter. By default no filtering is done.}
}
\value{
A vector of length equal to the number of paths in treesObj
}
\description{
turns a named vector of characters into a paths vector to be used with \code{\link{getAllCor}}
}

