% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordsfn.R
\name{click_select_foreground_branches}
\alias{click_select_foreground_branches}
\title{Interactive click-based function to select foreground branches showing convergent binary trait}
\usage{
click_select_foreground_branches(mastertre)
}
\arguments{
\item{mastertre.}{A phylo tree object defining the topology of all species}
}
\value{
A binary trait tree with branch lengths of 1 for selected foreground species and lengths of 0 for the rest
}
\description{
Interactive click-based function to select foreground branches showing convergent binary trait
}
