% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{combinePermData}
\alias{combinePermData}
\title{Combines batches of permulations}
\usage{
combinePermData(permdat1, permdat2, enrich = T)
}
\arguments{
\item{permdat1}{Batch of permulations output from \code{\link{getPermsContinuous}}}

\item{permdat2}{Batch of permulations output from \code{\link{getPermsContinuous}}}

\item{enrich}{Default T. Specifies if both `permdat1` and `permdat2` contain enrichment permulations}
}
\value{
Combined permulations
}
\description{
Combines batches of permulations
}
