% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr_diagnostics.R
\name{compareRateModels}
\alias{compareRateModels}
\title{Performs pairwise likelihood ratio comparisons on a list of rate models}
\usage{
compareRateModels(
  rate_models,
  treesObj,
  phenvals,
  nsims = 100,
  nested_only = FALSE,
  return_type = "pvals",
  ...
)
}
\arguments{
\item{rate_models}{a named list of rate models in matrix form}

\item{treesObj}{the trees object returned by readTrees}

\item{phenvals}{the named phenotype vector with names matching the tip labels on the tree}

\item{nsims}{the number of simulations to use in getLRnullDistribuutions when the rate models are not nested}

\item{nested_only}{a boolean indicating whether comparisons should only be made between nested models. If TRUE, this decreases runtime because simulations to get the null distribution are not performed.}

\item{return_type}{can be "pvals" or "ratios" indicating whether to return p values obtained from the likelihood ratios or the likelihood ratios themselves}

\item{...}{additional parameters for the castor function, fit_mk used to fit the transition matrix}
}
\value{
a square matrix of p-values or likelihood ratios for pairwise comparisons between rate models
}
\description{
Performs pairwise likelihood ratio comparisons on a list of rate models
}
