% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{edgeVars}
\alias{edgeVars}
\title{maps a vector of traits onto a reference tree}
\usage{
edgeVars(mastertree, tip.vals, metric = "diff", se.filter = -1,
  return.var = F)
}
\arguments{
\item{mastertree}{the tree species the topology of the output tree and the branch lengths are used to infer ancestral states}

\item{tip.vals}{the trait/phenotype/character value at the tip, \code{names(tip.vals)} should match some of the \code{mastertree$tip.label}, though a perfect match is not required}

\item{metric}{The metric used to translate node values into branch values. "Diff" takes the difference and makes the result phylogenetically independent. Other possible values are "mean" (the mean of the two values) and "last" the value of the most recent species on the branch. The last two options are not phylogenetically independent and downstream computations for those are not yet implemented}

\item{se.filter}{Will remove branch values that are not at least \code{se.filter*edge.se} away from 0 (where edge.se is the standard error in the estimate for the edge value). Only implemented for \code{metric="diff"}. Set \code{se.filter} to a positive value to filter. By default no filtering is done.}

\item{return.var}{Returns the variance instead of the mean. Useful for seeing which estimates have high confidence.}
}
\value{
A phylo tree with branch values computed from the input tip.values
}
\description{
maps a vector of traits onto a reference tree
}

