% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTreeFuncs.R
\name{estimatePhangornTree}
\alias{estimatePhangornTree}
\title{Estimate a ML tree from a given topology
Uses `pml` and `optim.pml` from the `phangorn` package to estimate the tree.}
\usage{
estimatePhangornTree(
  alnfile,
  treefile,
  submodel = "LG",
  type = "AA",
  format = "fasta",
  k = 4,
  ...
)
}
\arguments{
\item{alnfile}{The path to the sequence alignment file}

\item{treefile}{The path to the master tree file (whose topology will be used to generate the tree)}

\item{submodel}{Substitution model to be used to estimate the tree}

\item{type}{"AA" for amino acid or "DNA" for DNA}

\item{format}{Format of the alignment file (commonly used formats include fasta and phylip)}

\item{k}{Number of intervals in the discrete gamma distribution for `pml`}

\item{...}{Further arguments passed to `pml` or `optim.pml`}
}
\value{
A list:
`tree.opt` is the tree from the optimized output of `optim.pml`;
`results.opt` is the optimized output of `optim.pml`;
`results.init` is the initial results estimated by `pml`
}
\description{
Estimate a ML tree from a given topology
Uses `pml` and `optim.pml` from the `phangorn` package to estimate the tree.
}
\seealso{
\code{\link[phangorn]{phyDat}} for alignment formats,
 \code{\link[phangorn]{pml}} and \code{\link[phangorn]{optim.pml}} for tree estimation
}
