% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTreeFuncs.R
\name{estimatePhangornTreeAll}
\alias{estimatePhangornTreeAll}
\title{Estimates a ML trees from a given topology for a list of alignments. Wirtes a text file compatble with readTrees().
Uses `pml` and `optim.pml` from the `phangorn` package to estimate the tree.}
\usage{
estimatePhangornTreeAll(alnfiles = NULL, alndir = NULL,
  pattern = NULL, treefile, output.file = NULL, submodel = "LG",
  type = "AA", format = "fasta", k = 4, ...)
}
\arguments{
\item{alnfiles}{A character vector of paths to alignment files. Such as one produced by \code{\link{list.files}}. Either alnfiles or alndir must be supplied.}

\item{alndir}{The path to directories with alignment files. Trees will be named with the alignment file name after stripping the extension. Either alnfiles or alndir must be supplied.}

\item{pattern}{An optional regular expression for files in the alndir director. As in  ".*fasta".}

\item{treefile}{The path to the master tree file (whose topology will be used to generate the tree)}

\item{output.file}{The file where the output trees will be written. This file can be read with readTrees().}

\item{submodel}{Substitution model to be used to estimate the tree}

\item{type}{"AA" for amino acid or "DNA" for DNA}

\item{format}{Format of the alignment file (commonly used formats include fasta and phylip)}

\item{k}{Number of intervals in the discrete gamma distribution for `pml`}

\item{...}{Further arguments passed to `pml` or `optim.pml`}
}
\value{
A list:
`tree.opt` is the tree from the optimized output of `optim.pml`;
`results.opt` is the optimized output of `optim.pml`;
`results.init` is the initial results estimated by `pml`
}
\description{
Estimates a ML trees from a given topology for a list of alignments. Wirtes a text file compatble with readTrees().
Uses `pml` and `optim.pml` from the `phangorn` package to estimate the tree.
}
\seealso{
\code{\link[phangorn]{phyDat}} for alignment formats,
 \code{\link[phangorn]{pml}} and \code{\link[phangorn]{optim.pml}} for tree estimation
}
