% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentFuncs.R
\name{fastwilcoxGMTall}
\alias{fastwilcoxGMTall}
\title{Performs Wilcoxon Rank-Sum pathway enrichment on pathway genesets}
\usage{
fastwilcoxGMTall(vals, annotList, alternative = "two.sided", ...)
}
\arguments{
\item{vals}{The output from getStat run on RERconverge correlation functions (correlateWithContinuousPhenotype, correlateWithBinaryPhenotype, getAllCor)}

\item{annotList}{Pathway annotations formatted as a list of lists of annotation information}
}
\value{
A list of pathways with enrichment statistics
}
\description{
Performs Wilcoxon Rank-Sum pathway enrichment on pathway genesets
}
