% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr_diagnostics.R
\name{filterByCriteria}
\alias{filterByCriteria}
\title{Filters a list of rate models based on a boolean expression expressing filtering criteria}
\usage{
filterByCriteria(models, criteria)
}
\arguments{
\item{models}{a list of rate models}

\item{criteria}{a function that is passed to sapply specifying the desired criteria to filter by. It should take one argument, a rate model, and should be a boolean expression expressing criteria for the rate model.}
}
\value{
returns a subset of the list of rate models that match the criteria specified in the criteria function
}
\description{
Filters a list of rate models based on a boolean expression expressing filtering criteria
}
