% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{fixPseudoroot}
\alias{fixPseudoroot}
\title{pseudoroot trait tree to match the psuedoroot of trees from \code{\link{readTrees}}  if the trees are reconcilable}
\usage{
fixPseudoroot(tree, treesObj)
}
\arguments{
\item{tree}{A trait tree with branch lengths representing trait values}
}
\value{
A trait tree with the correct topology
}
\description{
pseudoroot trait tree to match the psuedoroot of trees from \code{\link{readTrees}}  if the trees are reconcilable
}
