% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{flat_to_ij}
\alias{flat_to_ij}
\title{Flat_to_ij}
\usage{
flat_to_ij(index, ncol)
}
\arguments{
\item{index}{A flast index}

\item{ncol}{The number of columns in the matrix}
}
\value{
A two column matrix of indeces corresponding to i and j
}
\description{
Takes a flat matrix index and returns it as two colums or row (i) and column (j) indeces
}
