% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{foreground2Paths}
\alias{foreground2Paths}
\title{Obtain a trees object where the foreground species have branch lengths of 1, and the rest 0}
\usage{
foreground2Paths(foreground, treesObj, plotTree = F)
}
\arguments{
\item{treesObj}{A treesObj created by \code{\link{readTrees}}}

\item{plotTree}{Plot a tree representation of the result}

\item{foreground.}{A character vector containing the foreground species}
}
\value{
A vector of length equal to the number of paths in treesObj
}
\description{
Obtain a trees object where the foreground species have branch lengths of 1, and the rest 0
}
