% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{foreground2Tree}
\alias{foreground2Tree}
\title{Creates a tree form a set of foreground species}
\usage{
foreground2Tree(foreground, treesObj, collapse2anc = T, plotTree = T,
  wholeClade = F)
}
\arguments{
\item{treesObj}{A treesObj created by \code{\link{readTrees}}}

\item{collapse2anc}{Put all the weight on the ancestral branch when the trait appears on a while clade}

\item{plotTree}{Plot a tree representation of the result}

\item{foreground.}{A character vector containing the foreground species}
}
\value{
A tree with edge.lengths representing phenotypic states
}
\description{
Creates a tree form a set of foreground species
}
