% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{generatePermulatedBinPhen}
\alias{generatePermulatedBinPhen}
\title{Produces binary permulations for a gene}
\usage{
generatePermulatedBinPhen(
  tree,
  numperms,
  trees,
  root_sp,
  fg_vec,
  sisters_list,
  pathvec,
  permmode = "cc",
  transition = "unidirectional"
)
}
\arguments{
\item{tree}{Tree of the gene of interest (if permmode="cc", set this as the masterTree in trees (i.e., the output from \code{\link{readTrees}}))}

\item{numperms}{An integer number of permulations}

\item{trees}{treesObj from \code{\link{readTrees}}}

\item{root_sp}{The species to root the tree on}

\item{fg_vec}{A vector containing the foreground species}

\item{sisters_list}{A list containing pairs of "sister species" in the foreground set (put NULL if empty)}

\item{pathvec}{A path vector generated from the real set of foreground animals}

\item{transition}{A character string indicating whether transitions between background and foreground branches
are "bidirectional" or "unidirectional" (no foreground to background transitions, the default)}
}
\value{
output.list a list containing the set of binary permulated trees
}
\description{
Produces binary permulations for a gene
}
