% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{generatePermulatedBinPhenSSMBatched}
\alias{generatePermulatedBinPhenSSMBatched}
\title{Produces binary SSM permulations for a list of genes}
\usage{
generatePermulatedBinPhenSSMBatched(
  trees_list,
  numperms,
  trees,
  root_sp,
  fg_vec,
  sisters_list,
  pathvec,
  permmode = "ssm"
)
}
\arguments{
\item{trees_list}{A list containing the trees of all genes of interest (formatted like trees in treesObj from \code{\link{readTrees}})}

\item{numperms}{An integer number of permulations}

\item{trees}{treesObj from \code{\link{readTrees}}}

\item{root_sp}{The species to root the tree on}

\item{fg_vec}{A vector containing the foreground species}

\item{sisters_list}{A list containing pairs of "sister species" in the foreground set (put NULL if empty)}

\item{pathvec}{A path vector generated from the real set of foreground animals}

\item{permmode}{Mode of binary permulation (Must be "ssm" or "ssmLegacy" for this function; default "ssm")}
}
\value{
simPhenoList A list containing binary permulated trees for each gene
}
\description{
Produces binary SSM permulations for a list of genes
}
