% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{getBinaryPermulationInputsFromTree}
\alias{getBinaryPermulationInputsFromTree}
\title{Produces the phenotype inputs for binary permulation functions}
\usage{
getBinaryPermulationInputsFromTree(fgTree)
}
\arguments{
\item{fgTree}{A binary phenotype tree containing the phenotype values (branch length = 1 for foreground species, 0 for background species)}
}
\value{
out A list containing fg_vec and sisters_list for binary permulation functions
}
\description{
Produces the phenotype inputs for binary permulation functions
}
