% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{getEnrichPermsContinuous}
\alias{getEnrichPermsContinuous}
\title{Performs enrichment statistic permulations using existing gene correlation permulations}
\usage{
getEnrichPermsContinuous(corperms, realenrich, annotlist)
}
\arguments{
\item{corperms}{Gene correlation permulations from \code{\link{getPermsContinuous}}}

\item{realenrich}{Pathway enrichment results using observed phenotype obtained from `correlateWithContinuousPhenotype`}

\item{annotlist}{Pathway annotations}
}
\value{
Full null permulation statistics and p-values for gene correlations and pathway enrichment
}
\description{
Performs enrichment statistic permulations using existing gene correlation permulations
}
