% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{getForegroundInfoClades}
\alias{getForegroundInfoClades}
\title{Generates a binary phenotype tree and foreground clades information using the list of tip foreground animals, the presence of foreground common ancestors, and their phylogenetic relationships}
\usage{
getForegroundInfoClades(
  fg_vec,
  sisters_list = NULL,
  trees,
  plotTree = T,
  useSpecies = NULL,
  transition = "unidirectional"
)
}
\arguments{
\item{fg_vec}{A vector containing the tip foreground species}

\item{sisters_list}{A list containing pairs of "sister species" in the foreground set (put NULL if empty)}

\item{trees}{treesObj from \code{\link{readTrees}}}

\item{plotTree}{A boolean indicator for plotting the output tree (default=FALSE)}

\item{useSpecies}{An array containing the tip labels in the output tree}

\item{transition}{A character string indicating whether transitions between background and foreground branches
are "bidirectional" or "unidirectional" (no foreground to background transitions, the default)}
}
\value{
output.list A list containing 1) "tree" = a binary phenotype tree corresponding to the input information, 2) "fg.sisters.table" = a table containing all sister species in the foreground set
}
\description{
Generates a binary phenotype tree and foreground clades information using the list of tip foreground animals, the presence of foreground common ancestors, and their phylogenetic relationships
}
