% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr_diagnostics.R
\name{getLikelihoodRatio}
\alias{getLikelihoodRatio}
\title{compute the likelihood ratio from the log likeilhoods of two transition matrices}
\usage{
getLikelihoodRatio(limited, general)
}
\arguments{
\item{limited}{The log likelihood of the simpler model}

\item{general}{The log likelihood of the more complex model}
}
\value{
the likelihood ratio between two rate models
}
\description{
compute the likelihood ratio from the log likeilhoods of two transition matrices
}
