% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr_diagnostics.R
\name{getLRnullDistribution}
\alias{getLRnullDistribution}
\title{Returns the p value of a likelihood ratio based on the simulated null distribution}
\usage{
getLRnullDistribution(
  LR,
  tree,
  Q,
  null_rm,
  alt_rm,
  nsims = 100,
  plot = FALSE,
  ...
)
}
\arguments{
\item{LR}{the likelihood ratio between two rate models}

\item{tree}{the phylogenetic tree to obtain the null distribution on}

\item{Q}{the transition matrix fit on the null rate model}

\item{null_rm}{the null rate model (the simpler rate model)}

\item{alt_rm}{the alternative rate model (the more complex rate model)}

\item{nsims}{the number of simulations to use generate the null distribution}

\item{plot}{a boolean expressing whether or not to plot a histogram of likeilhood ratios representing the null distribution}
}
\value{
the p value of a given likeihood ratio based on the null distribution that is generated
}
\description{
Returns the p value of a likelihood ratio based on the simulated null distribution
}
