% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr_diagnostics.R
\name{getMatrixFromAbbr}
\alias{getMatrixFromAbbr}
\title{Gets the matrix form of a rate model from its abbreviation}
\usage{
getMatrixFromAbbr(abbr, Nstates)
}
\arguments{
\item{abbr}{The abbreviation for the rate model. Can be one of "ER", "ARD", or "SYM"}

\item{Nstates}{The number of phenotype states to include in the rate model}
}
\value{
a rate model as an Nstates x Nstates matrix
}
\description{
Gets the matrix form of a rate model from its abbreviation
}
