% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{getPerms}
\alias{getPerms}
\title{Calculates permuted correlation and enrichment statistics}
\usage{
getPerms(numperms, traitvec, RERmat, annotlist, treetop, trees,
  type = "simperm", winR = NULL, winT = NULL)
}
\arguments{
\item{numperms}{An integer number of permulations}

\item{traitvec}{A named phenotype vector}

\item{RERmat}{An RER matrix calculated using \code{\link{getAllResiduals}}}

\item{annotlist}{Pathway annotations}

\item{treetop}{A rooted, fully dichotomous tree derived from the treesObj master tree from \code{\link{readTrees}}.  Must not contain species not in traitvec}

\item{trees}{treesObj from \code{\link{readTrees}}}

\item{type}{One of "simperm", "sim", or "perm" for permulations, simulations, or permutations, respectively}

\item{winR}{Integer winzorization value for RERs}

\item{winT}{Integer winzorization value for trait}
}
\value{
A list object with enrichment statistics, correlation p-val, rho, and correlation effect size
}
\description{
Calculates permuted correlation and enrichment statistics
}
\note{
winsorize is in terms of number of observations at each end, NOT quantiles
}
