% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{getPermsBinaryFudged}
\alias{getPermsBinaryFudged}
\title{Calculates permuted correlation and enrichment statistics for binary phenotype (does not enforce matching structure of sister species, but maintains internal & tips foreground counts within a fudge of the original tree)}
\usage{
getPermsBinaryFudged(
  fgdspecs,
  RERs,
  trees,
  useSpecies,
  ntrees,
  root,
  fudge = 5,
  cors,
  phenvec,
  transition = "unidirectional"
)
}
\arguments{
\item{fgdspecs}{A vector containing the tip foreground species}

\item{RERs}{An RER matrix calculated using \code{\link{getAllResiduals}}.}

\item{trees}{treesObj from \code{\link{readTrees}}}

\item{useSpecies}{A vector of species to include}

\item{ntrees}{An integer number of permulations}

\item{root}{The species to root the tree on}

\item{fudge}{The number up to which the permulated tree can differ in total foreground species}

\item{cors}{The gene correlation results from \code{\link{correlateWithBinaryPhenotype}}}

\item{phenvec}{A named vector of the phenotype}

\item{transition}{A character string indicating whether transitions between background and foreground branches
are "bidirectional" or "unidirectional" (no foreground to background transitions, the default)}
}
\value{
A list object with enrichment statistics, correlation p-val, rho, and correlation effect size
}
\description{
Calculates permuted correlation and enrichment statistics for binary phenotype (does not enforce matching structure of sister species, but maintains internal & tips foreground counts within a fudge of the original tree)
}
