% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{getPermsContinuousExtantOnly}
\alias{getPermsContinuousExtantOnly}
\title{Calculates permuted correlation and enrichment statistics for an extant only analysis}
\usage{
getPermsContinuousExtantOnly(
  numperms,
  traitvec,
  RERmat,
  annotlist,
  trees,
  mastertree,
  calculateenrich = T,
  type = "simperm",
  winR = 3,
  winT = 3,
  method = "p",
  min.pos = 0
)
}
\arguments{
\item{numperms}{An integer number of permulations}

\item{traitvec}{A named phenotype vector}

\item{RERmat}{An RER matrix calculated using \code{\link{getAllResiduals}}}

\item{annotlist}{Pathway annotations}

\item{trees}{treesObj from \code{\link{readTrees}}}

\item{mastertree}{A rooted, fully dichotomous tree derived from the treesObj master tree from \code{\link{readTrees}}.  Must not contain species not in traitvec}

\item{calculateenrich}{A boolean variable indicating if null permulation p-values for enrichment statistics}

\item{type}{One of "simperm", "sim", or "perm" for permulations, simulations, or permutations, respectively}

\item{winR}{Integer winzorization value for RERmat}

\item{winT}{Integer winzorization value for trait}

\item{method}{statistical method to use for correlations}

\item{min.pos}{minimum foreground species - should be set to 0}
}
\value{
A list object with enrichment statistics, correlation p-val, rho, and correlation effect size
}
\description{
Calculates permuted correlation and enrichment statistics for an extant only analysis
}
\note{
winsorize is in terms of number of observations at each end, NOT quantiles
}
