% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{getStat}
\alias{getStat}
\title{Calculates Rho-signed negative log-base-ten p-value for use in enrichment functions}
\usage{
getStat(res)
}
\arguments{
\item{res}{The output from RERconverge correlation functions (correlateWithContinuousPhenotype, correlateWithBinaryPhenotype, getAllCor)}
}
\value{
A dataframe of Rho-signed negative log-base-ten p-values for all genes, NAs removed
}
\description{
Calculates Rho-signed negative log-base-ten p-value for use in enrichment functions
}
