% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{getStatesFromPhenTree}
\alias{getStatesFromPhenTree}
\title{Returns a list of states in node order from a phenotype tree}
\usage{
getStatesFromPhenTree(phenTree, root_state, tipvals)
}
\arguments{
\item{phenTree}{The phenotype tree returned by char2TreeCategorical}

\item{root_state}{The state of the root node, this is not stored along the edge lengths of the tree}

\item{tipvals}{The states of the tips of the tree in order of phenTree$tip.label}
}
\value{
Returns a vector of the category states of species in order of the nodes in the tree
}
\description{
Returns a list of states in node order from a phenotype tree
}
