% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{kwdunn.test}
\alias{kwdunn.test}
\title{A sped up version of the Kruskal Wallis/Dunn Test}
\usage{
kwdunn.test(x, g, ncategories)
}
\description{
A sped up version of the Kruskal Wallis/Dunn Test
}
\keyword{internal}
