% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{mapIndex}
\alias{mapIndex}
\title{Map index}
\usage{
mapIndex(treesObj1, treesObj2)
}
\arguments{
\item{treesObj1}{The first trees object}

\item{treesObj2}{The second trees object}
}
\value{
The index of the paths matrix columns in treesObj1 inside treesObj2
}
\description{
Creates a map from the paths in one tree object to the paths in the other object
Can be used to put old style and new style trees objects into the same order
}
