% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{matchAllNodesClades}
\alias{matchAllNodesClades}
\title{Calculates the clade mappings between the gene tree and the master tree (with the complete topology)}
\usage{
matchAllNodesClades(gene_tree, treesObj)
}
\arguments{
\item{gene_tree}{A binary phenotype tree of a gene}

\item{treesObj}{treesObj from \code{\link{readTrees}}}
}
\value{
output.map A list containing a dataframe of clades mapping
}
\description{
Calculates the clade mappings between the gene tree and the master tree (with the complete topology)
}
