% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFuncs.R
\name{namePathsWSpecies}
\alias{namePathsWSpecies}
\title{Extract a vector of names corresponding to paths on the given tree}
\usage{
namePathsWSpecies(masterTree)
}
\arguments{
\item{masterTree}{A tree object.}

}
\value{
A vector of names corresponding to paths on the given tree.
}
\description{
Extract a vector of names corresponding to paths on the given tree
}
