% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{permpvalenrich}
\alias{permpvalenrich}
\title{Calculates enrichment permutation pvals from output of \code{\link{getPermsContinuous}}}
\usage{
permpvalenrich(realenrich, permvals)
}
\arguments{
\item{realenrich}{Real enrichment statistics from \code{\link{fastwilcoxGMTall}}}

\item{permvals}{output from \code{\link{getPermsContinuous}}}
}
\value{
A list object with vectors of permulation p-values
}
\description{
Calculates enrichment permutation pvals from output of \code{\link{getPermsContinuous}}
}
