% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{permutevec}
\alias{permutevec}
\title{Generates a permuted continuous phenotype given an observed continuous phenotype}
\usage{
permutevec(namedvec)
}
\arguments{
\item{namedvec}{A named numeric vector with phenotype values for each speices}
}
\value{
A vector with permuted phenotype values
}
\description{
Generates a permuted continuous phenotype given an observed continuous phenotype
}
