% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFuncs.R
\name{plotRers}
\alias{plotRers}
\title{Plot the residuals reflecting the relative evolutionary rates (RERs) of a gene across species present in the gene tree}
\usage{
plotRers(rermat = NULL, index = NULL, phenv = NULL, rers = NULL,
  method = "s", plot = 1, xextend = 0.2, sortrers = F)
}
\arguments{
\item{rermat.}{A residual matrix, output of the getAllResiduals() function}

\item{index.}{A character denoting the name of gene, or a numeric value corresponding to the gene's row index in the residuals matrix}

\item{phenv.}{A phenotype vector returned by \code{\link{tree2Paths}} or \code{\link{foreground2Paths}}}
}
\value{
A plot of the RERs with foreground species labelled in red, and the rest in blue
}
\description{
Plot the residuals reflecting the relative evolutionary rates (RERs) of a gene across species present in the gene tree
}
