% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFuncs.R
\name{plotTreeHighlightBranches}
\alias{plotTreeHighlightBranches}
\title{Plot the provided tree, (optionally) rerooted, with specified branches highlighted}
\usage{
plotTreeHighlightBranches(
  tree,
  outgroup = NULL,
  hlspecies,
  hlcols = NULL,
  main = ""
)

plotTreeHighlightBranches(
  tree,
  outgroup = NULL,
  hlspecies,
  hlcols = NULL,
  main = ""
)
}
\arguments{
\item{hlspecies}{the species to highlight}

\item{hlcols}{the colors to highlight those species with}

\item{main}{the name to give the plot}

\item{tree.}{A tree object.}

\item{outgroup.}{A vector of species to use to root the tree. If not provided, the tree remains unrooted.}

\item{hlspecies.}{A vector of species whose terminal branches to highlight, or a vector of branch numbers within the tree.}

\item{hlcols.}{Colors to use in highlighting the branches. If not specified, will use default R colors.}

\item{main.}{Main text for plot.}

\item{useGG}{option to creat plot with ggtree. Improves plot readability for most data sets.}
}
\value{
A plot of the the (optionally rerooted) tree, with branches highlighted.
}
\description{
Plot the provided tree, (optionally) rerooted, with specified branches highlighted
}
