% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{pruneTree}
\alias{pruneTree}
\title{wrapper around \code{\link[ape]{drop.tip}}}
\usage{
pruneTree(tree, tip.names)
}
\arguments{
\item{tree}{a "phylo" tree}

\item{tip.names}{The tip names to keep in the tree}
}
\value{
A new pruned tree
}
\description{
wrapper around \code{\link[ape]{drop.tip}}
}

