% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentFuncs.R
\name{read.gmt}
\alias{read.gmt}
\title{Reads pathway data from a gmt-formatted file from GSEA-MSigDB}
\usage{
read.gmt(filename)
}
\arguments{
\item{filename}{The gmt-formatted pathway file}
}
\value{
A list of pathway information with pathway names and genesets
}
\description{
Reads pathway data from a gmt-formatted file from GSEA-MSigDB
}
