% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFuncs.R
\name{returnRersAsNewickStrings}
\alias{returnRersAsNewickStrings}
\title{Produce a vector of newick strings representing gene trees where the edge lengths correspond to RERs}
\usage{
returnRersAsNewickStrings(treesObj, rermat)
}
\arguments{
\item{treesObj.}{A treesObj created by \code{\link{readTrees}}}

\item{rermat.}{A residual matrix, output of the getAllResiduals() function}
}
\value{
A named character vector of newick strings, one per gene, representing RERs as edge lengths'
}
\description{
Produce a vector of newick strings representing gene trees where the edge lengths correspond to RERs
}
