% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFuncs.R
\name{returnRersAsTree}
\alias{returnRersAsTree}
\title{Produce a gene tree with branch lengths representing RERs and (optionally)
display these RERs as branch labels}
\usage{
returnRersAsTree(treesObj, rermat, index, phenv = NULL, rer.cex = 0.7,
  tip.cex = 0.7, nalab = "NA", plot = T)
}
\arguments{
\item{treesObj.}{A treesObj created by \code{\link{readTrees}}}

\item{rermat.}{A residual matrix, output of the getAllResiduals() function}

\item{index.}{A character denoting the name of gene, or a numeric value corresponding to the gene's row index in the residuals matrix}

\item{phenv.}{A phenotype vector returned by \code{\link{tree2Paths}} or \code{\link{foreground2Paths}}}

\item{rer.cex.}{Numeric expansion for RER labels}

\item{tip.cex.}{Numeric expansion for tip labels}

\item{nalab.}{Label given to any NA RERs}

\item{plot.}{Whether to produce a plot displaying the RERs on the gene tree}
}
\value{
An object of class "phylo" with edge lengths representing RERs for the given gene

If plot = TRUE, also displays a plot of the gene tree with edges labeled with RERs
}
\description{
Produce a gene tree with branch lengths representing RERs and (optionally)
display these RERs as branch labels
}
