% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFuncs.R
\name{returnRersAsTreesAll}
\alias{returnRersAsTreesAll}
\title{Produce a multiPhylo object of all gene trees with branch lengths representing RERs
for each gene}
\usage{
returnRersAsTreesAll(treesObj, rermat)
}
\arguments{
\item{treesObj.}{A treesObj created by \code{\link{readTrees}}}

\item{rermat.}{A residual matrix, output of the getAllResiduals() function}
}
\value{
An object of class "multiPhylo" of named gene trees with edge lengths
representing RERs for the given gene
}
\description{
Produce a multiPhylo object of all gene trees with branch lengths representing RERs
for each gene
}
