% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{Root}
\alias{Root}
\title{the function root from ape modified to return additional information about node matching and which edges are inverted
ape citation: Paradis E. & Schliep K. 2019. ape 5.0: an environment for modern phylogenetics and evolutionary analyses in R.Bioinformatics 35: 526-528.}
\usage{
Root(
  phy,
  outgroup,
  node = NULL,
  resolve.root = FALSE,
  interactive = FALSE,
  edgelabel = FALSE,
  ...
)
}
\description{
the function root from ape modified to return additional information about node matching and which edges are inverted
ape citation: Paradis E. & Schliep K. 2019. ape 5.0: an environment for modern phylogenetics and evolutionary analyses in R.Bioinformatics 35: 526-528.
}
\keyword{internal}
