% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{simBinPheno}
\alias{simBinPheno}
\title{Generates a permulated phylogenetic tree with specified number of foreground branches.  User may specify the number of foreground branches that are internal branches.}
\usage{
simBinPheno(trees, root, phenvec, fgnum = NULL, internal = 0, drop = NULL)
}
\arguments{
\item{trees}{treesObj output from \code{\link{readTrees}}}

\item{root}{Species on which to root the master tree}

\item{phenvec}{Named vector of 1's and 0's representing phenotype values for each species}

\item{fgnum}{Total number of foreground species - only required if internal foreground branches are required}

\item{internal}{Number of foreground species that should be internal branches - only required if internal foreground branches are required}

\item{drop}{Character vector (or single character variable) of species names to be removed from the master tree (such as species in trees but not in phenotype vector)}
}
\value{
A tree with permulated phenotype values
}
\description{
Generates a permulated phylogenetic tree with specified number of foreground branches.  User may specify the number of foreground branches that are internal branches.
}
