% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{simBinPhenoCCmidpoint}
\alias{simBinPhenoCCmidpoint}
\title{Produces one CC binary permulation for a gene using midpoint rooting}
\usage{
simBinPhenoCCmidpoint(
  trees,
  mastertree,
  fg_vec,
  sisters_list = NULL,
  pathvec,
  plotTreeBool = F
)
}
\arguments{
\item{trees}{treesObj from \code{\link{readTrees}}}

\item{mastertree}{A rooted, fully dichotomous tree derived from the treesObj master tree from \code{\link{readTrees}}.}

\item{fg_vec}{a vector containing the foreground species}

\item{sisters_list}{A list containing pairs of "sister species" in the foreground set (put NULL if empty)}

\item{pathvec}{A path vector generated from the real set of foreground animals}

\item{plotTreeBool}{Boolean indicator for plotting the output tree (default=FALSE)}
}
\value{
A CC binary permulated tree
}
\description{
Produces one CC binary permulation for a gene using midpoint rooting
}
