% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{simBinPhenoRank}
\alias{simBinPhenoRank}
\title{Produces one binary permulation based on ranking of simulated branch lengths}
\usage{
simBinPhenoRank(trees, root_sp, fg_vec, sisters_list = NULL, plotTreeBool = F)
}
\arguments{
\item{trees}{treesObj from \code{\link{readTrees}}}

\item{root_sp}{The species to root the tree on}

\item{fg_vec}{a vector containing the foreground species}

\item{sisters_list}{A list containing pairs of "sister species" in the foreground set (put NULL if empty)}

\item{plotTreeBool}{Boolean indicator for plotting the output tree (default=FALSE)}
}
\value{
A binary permulated tree
}
\description{
Produces one binary permulation based on ranking of simulated branch lengths
}
