% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{simBinPhenoSSMLegacy}
\alias{simBinPhenoSSMLegacy}
\title{Produces one SSM binary permulation for a gene}
\usage{
simBinPhenoSSMLegacy(
  tree,
  trees,
  root_sp,
  fg_vec,
  sisters_list = NULL,
  pathvec,
  plotTreeBool = F
)
}
\arguments{
\item{tree}{Tree of the gene of interest}

\item{trees}{treesObj from \code{\link{readTrees}}}

\item{root_sp}{The species to root the tree on}

\item{fg_vec}{A vector containing the foreground species}

\item{sisters_list}{A list containing pairs of "sister species" in the foreground set (put NULL if empty)}

\item{pathvec}{A path vector generated from the real set of foreground animals}

\item{plotTreeBool}{Boolean indicator for plotting the output tree (default=FALSE)}
}
\value{
A SSM binary permulated tree
}
\description{
Produces one SSM binary permulation for a gene
}
