% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{simpermvec}
\alias{simpermvec}
\title{Generates a permulated continuous phenotype given an observed continuous phenotype and a phylogeny}
\usage{
simpermvec(namedvec, treewithbranchlengths)
}
\arguments{
\item{namedvec}{A named numeric vector with phenotype values for each speices}

\item{treewithbranchlengths}{A rooted phylogenetic tree with the same species as namedvec and branch lengths representing average evolutionary rate. The master tree from \code{\link{readTrees}} may be rooted and used for this parameter.}
}
\value{
A vector with permulated phenotype values
}
\description{
Generates a permulated continuous phenotype given an observed continuous phenotype and a phylogeny
}
