% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{simulatevec}
\alias{simulatevec}
\title{Generates a simulated continuous phenotype given an observed continuous phenotype and a phylogeny}
\usage{
simulatevec(namedvec, treewithbranchlengths)
}
\arguments{
\item{namedvec}{A named numeric vector with phenotype values for each speices}

\item{treewithbranchlengths}{A rooted phylogenetic tree with the same species as namedvec and branch lengths representing average evolutionary rate. The master tree from \code{\link{readTrees}} may be rooted and used for this parameter.}
}
\value{
A vector with simulated phenotype values
}
\description{
Generates a simulated continuous phenotype given an observed continuous phenotype and a phylogeny
}
