% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFuncs.R
\name{translateNames}
\alias{translateNames}
\title{setup step to change names of plotRers points}
\usage{
translateNames(from, to, row)
}
\description{
setup step to change names of plotRers points
}
\keyword{internal}
