% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{tree2Paths}
\alias{tree2Paths}
\title{Generates a phenotype paths vector matching thre treesObject from a tree where branches specify phenotypes}
\usage{
tree2Paths(tree, treesObj, binarize = T)
}
\arguments{
\item{tree}{A phenotype tree, with branch length encoding a phenotype.}

\item{treesObj}{A treesObj created by \code{\link{readTrees}}}

\item{binarize}{Force binary path representation. Sets all positive path values to 1. Useful if the tree has ancestral nodes with non-zero branches, otherwise values are simply added along branches.}
}
\value{
A vector of length equal to the number of paths in treesObj
}
\description{
Generates a phenotype paths vector matching thre treesObject from a tree where branches specify phenotypes
}
