% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{tree2Paths_map}
\alias{tree2Paths_map}
\title{A modification of the tree2Paths function that takes in pre-calculated mappings}
\usage{
tree2Paths_map(tree, map, treesObj, binarize = NULL, useSpecies = NULL)
}
\arguments{
\item{tree}{the input tree to be converted into paths}

\item{map}{a dataframe containing clade mappings corresponding to the tree}

\item{treesObj}{a multiphylo object containing all the gene trees and the master tree}
}
\description{
A modification of the tree2Paths function that takes in pre-calculated mappings
}
