% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{tree2PathsClades}
\alias{tree2PathsClades}
\title{A modification of the tree2Paths function that takes in pre-calculated mappings}
\usage{
tree2PathsClades(tree, trees)
}
\arguments{
\item{tree}{the input tree to be converted into paths}

\item{trees}{treesObj from \code{\link{readTrees}}}
}
\description{
A modification of the tree2Paths function that takes in pre-calculated mappings
}
