% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFuncs.R
\name{treePlotNew}
\alias{treePlotNew}
\title{Plot `tree` with branch labels colored in a heatmap based on values in `vals`}
\usage{
treePlotNew(
  tree,
  vals = NULL,
  rank = F,
  nlevels = 9,
  type = "c",
  col = NULL,
  maintitle = NULL,
  useedge = F,
  doreroot = F,
  rerootby = NULL,
  useSpecies = NULL,
  species.names = NULL,
  speclist1 = NULL,
  speclist2 = NULL,
  aligntip = F,
  colpan1 = rgb(0, 119, 187, maxColorValue = 255),
  colpan2 = rgb(204, 51, 17, maxColorValue = 255),
  colpanmid = rgb(187, 187, 187, maxColorValue = 255),
  plotspecies = NULL,
  edgetype = NULL,
  textsize = 0.6,
  colbarlab = "",
  splist2sym = "psi",
  dolegend = T,
  nacol = rgb(0, 0, 0),
  figwid = 10,
  ...
)
}
\arguments{
\item{tree.}{A phylo object, used for topology}

\item{vals.}{Values to use for heatmap to color branches, in same order as edges of tree}

\item{rank.}{Whether to plot ranks instead of values}

\item{nlevels.}{How many colors to use in the heatmap}

\item{type.}{Type for plot.phylo}

\item{col.}{Vector of user-defined colors}

\item{maintitle.}{Main title label}

\item{useedge.}{Whether to use edge lengths in `tree` for plotting}

\item{doreroot.}{Whether to re-root the tree before  plotting}

\item{rerootby.}{If re-rooting, what to use to root the tree (verify by checking against unrooted plot)}

\item{useSpecies.}{A vector of species to include in the plot (verify by checking against full set)}

\item{species.names.}{Data fram for converting names in `tree` to names to be plotted
(row names should be tip labels of tree, and first column should contain the
corresponding desired tip labels)}

\item{speclist1.}{A vector of tip labels to highlight in bold, blue text}

\item{speclist2.}{A vector of tip labels to which to add an asterisk}

\item{aligntip.}{Whether to align tip labels (default FALSE)}

\item{colpan1.}{Color for lowest value in heatmap (default blue)}

\item{colpan2.}{Color for highest value in heatmap (default rose)}

\item{colpanmid.}{Color for middle value in heatmap (default gray)}

\item{plotspecies.}{A vector of tip labels to display on the tree (the remainder will be masked,
but the corresponding tips will be plotted on the tree)}

\item{edgetype.}{Vector of line type for edges.}

\item{textsize.}{cex value for tip labels (default 0.6)}

\item{colbarlab.}{Label for the color bar legend}

\item{splist2sym.}{A value within species names to display as a symbol}

\item{dolegend.}{Whether to display the heatmap legend.}

\item{nacol.}{Color to display for any edges with length NA}

\item{figwid.}{Adjust x limits of plot.phylo by 1/figwid. May be related to figure width
(requires some optimization)}

\item{....}{further arguments to be passed to `plot` or to `plot.phylo`}
}
\value{
Plots a cladogram of `tree` with branch colors determined by `vals`
}
\description{
Plot `tree` with branch labels colored in a heatmap based on values in `vals`
}
