% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFuncs.R
\name{treePlotRers}
\alias{treePlotRers}
\title{Plot a cladogram with RERs shown as either labels (type="label") or a color
heatmap along the branches (type="color")
Wraps around \code{\link{returnRersAsTree}} or \code{\link{treePlotNew}}, respectively}
\usage{
treePlotRers(treesObj, rermat = NULL, index = NULL, type = c("label",
  "color"), phenv = NULL, ...)
}
\arguments{
\item{treesObj.}{A treesObj created by \code{\link{readTrees}}}

\item{rermat.}{A residual matrix, output of the getAllResiduals() function}

\item{index.}{A character denoting the name of gene, or a numeric value
corresponding to the gene's row index in the residuals matrix}

\item{type.}{Whether to display RERs as branch labels ('label') or a heatmap ('color')}

\item{phenv.}{A phenotype vector returned by \code{\link{tree2Paths}} or \code{\link{foreground2Paths}}}

\item{....}{Additional parameters to be passed to \code{\link{returnRersAsTree}}
or \code{\link{treePlotNew}}}
}
\value{
Plots a cladogram of the master tree with RERs displayed as branch labels or colors
}
\description{
Plot a cladogram with RERs shown as either labels (type="label") or a color
heatmap along the branches (type="color")
Wraps around \code{\link{returnRersAsTree}} or \code{\link{treePlotNew}}, respectively
}
