% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr_diagnostics.R
\name{visCompareTwoRateModels}
\alias{visCompareTwoRateModels}
\title{Visually compare the results of two rate models by plotting the differences between them on the phylogenetic tree
Relative entropy is calculated with A as the "target" and B as the "estimate"
Switching the order of rate model A and rate model B DOES give different relative entropies (relative entropy is not symmetric)}
\usage{
visCompareTwoRateModels(
  A,
  B,
  treesObj,
  phenvals,
  mode = "entropy",
  cex = 0.5,
  ...
)
}
\arguments{
\item{A}{the first rate model}

\item{B}{teh second rate model}

\item{treesObj}{the trees object returned by readTrees}

\item{phenvals}{the named phenotype vector with names matching the tip labels in the trees object}

\item{mode}{indicates how to represent differences between ancestral reconstructions. When mode is "entropy" the relative entropies between ancestral likelihoods at each node are plotted along the branches of the trees. When mode is "match", nodes are colored red when state assignments are different and red when they are the same.}

\item{cex}{a graphical argument representing the size of node labels and tip labels}
}
\value{
a list of relative entropies or state assignemtns and ancestral likelihoods under each rate model
}
\description{
Visually compare the results of two rate models by plotting the differences between them on the phylogenetic tree
Relative entropy is calculated with A as the "target" and B as the "estimate"
Switching the order of rate model A and rate model B DOES give different relative entropies (relative entropy is not symmetric)
}
