% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{adaptivePermulation}
\alias{adaptivePermulation}
\title{Adaptively calculates permulation p-value for a single element}
\usage{
adaptivePermulation(rer, permulated_foregrounds, observed_stats, alpha = 0.05)
}
\arguments{
\item{rer}{a row matrix for the given element (e.g., a row from the matrix output from \code{\link{getAllResiduals}})}

\item{permulated_foregrounds}{a list containing n sets of permulated foreground species names}

\item{observed_stats}{computed statistics for the observed trait (e.g., output from \code{\link{getAllCor}}, \code{\link{correlateWithBinaryPhenotype}}, or \code{\link{correlateWithContinuousPhenotype}})}

\item{alpha}{the significance level to control (default = 0.05)}
}
\value{
permPval permulation (empirical) p-value
}
\description{
Adaptively calculates permulation p-value for a single element
}
