% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{calculatePermulatedPaths_apply}
\alias{calculatePermulatedPaths_apply}
\title{Calculates the paths for all permulated trees for a list of genes}
\usage{
calculatePermulatedPaths_apply(permulated.trees.list, map.list, treesObj)
}
\arguments{
\item{permulated.trees.list}{A nested list of permulated phenotype trees for all the genes. The top layer lists the genes, and the nested layer lists the permulated binary trees (phylo objects) for each gene}

\item{map.list}{A list of maps corresponding to the genes listed in permulated.trees.list, the output of matchAllNodesClades}

\item{treesObj}{treesObj from \code{\link{readTrees}}}
}
\value{
permulated.paths.list A nested list of permulated paths corresponding to the nested list of permulated phenotype trees
}
\description{
Calculates the paths for all permulated trees for a list of genes
}
