% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{char2Paths}
\alias{char2Paths}
\title{turns a named vector of characters into a paths vector to be used with \code{\link{getAllCor}}}
\usage{
char2Paths(tip.vals, treesObj, altMasterTree = NULL, metric = "diff",
  se.filter = -1)
}
\value{
A vector of length equal to the number of paths in treesObj
}
\description{
turns a named vector of characters into a paths vector to be used with \code{\link{getAllCor}}
}
