% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{char2TreeCategorical}
\alias{char2TreeCategorical}
\title{Creates a categorical trait tree from a set of tip species.}
\usage{
char2TreeCategorical(
  tipvals,
  treesObj,
  useSpecies = NULL,
  use_rooted = FALSE,
  outgroup = NULL,
  model = "ER",
  plot = FALSE
)
}
\arguments{
\item{tipvals}{the trait/phenotype/character value at the tip, \code{names(tip.vals)} should match some of the \code{mastertree$tip.label}, though a perfect match is not required}

\item{treesObj}{A treesObj created by \code{\link{readTrees}}}

\item{useSpecies}{Give only a subset of the species to use for ancestral state reconstruction
(e.g., only those species for which the trait can be reliably determined).}

\item{use_rooted}{Root the tree first to use the phytools function ace. Otherwise leaves the tree unrooted and uses the castor function asr_mk_model.}

\item{outgroup}{If use_rooted is true, specifies what to use as the outgroup in order to root the tree}

\item{model}{Specifies what rate model to use}

\item{plot}{Plots a phenotype tree}
}
\value{
A tree with edge.lengths representing phenotypic states
}
\description{
Creates a categorical trait tree from a set of tip species.
}
