% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{correlateWithBinaryPhenotype}
\alias{correlateWithBinaryPhenotype}
\title{Computes the association statistics between RER from \code{\link{getAllResiduals}} and a phenotype paths vector for a binary phenotype made with \code{\link{tree2Paths}}}
\usage{
correlateWithBinaryPhenotype(
  RERmat,
  charP,
  min.sp = 10,
  min.pos = 2,
  weighted = "auto",
  winsorizeRER = NULL,
  winsorizetrait = NULL,
  bootstrap = F,
  bootn = 1000
)
}
\arguments{
\item{RERmat}{RER matrix returned by \code{\link{getAllResiduals}}}

\item{charP}{phenotype vector returned by \code{\link{tree2Paths}} or \code{\link{char2Paths}}}

\item{min.sp}{Minimum number of species that must be present for a gene}

\item{min.pos}{Minimum number of species that must be present in the foreground (non-zero phenotype values)}

\item{weighted}{perform weighted correlation. This option turns on weighted correlation that uses the weights computed by \code{\link{foreground2Tree}(wholeClade=T)}. This setting will treat each clade a single observation for the purpose of p-value estimation. The function will guess automatically if the charP vector is of "weighted" type and there should be not need to set this parameter.}

\item{winsorizeRER}{Winsorize RER values before computing Pearson correlation. winsorizeRER=3 will set the 3 most extreme RER values at each end of each row to the 4th most extreme value.}

\item{winsorizetrait}{Winsorize phenotype vector values before computing Pearson correlation. winsorizeRER=3 will set the 3 most extreme RER values at each end of each row to the 4th most extreme value.}

\item{bootstrap}{toggle bootstrapping (for weighted pearson correlation)}

\item{bootn}{number of runs to use when bootstrapping. Will be ignored if bootstrap is false.}
}
\description{
Computes the association statistics between RER from \code{\link{getAllResiduals}} and a phenotype paths vector for a binary phenotype made with \code{\link{tree2Paths}}
}
