% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{correlateWithCategoricalPhenotype}
\alias{correlateWithCategoricalPhenotype}
\title{Computes the association statistics between RER from \code{\link{getAllResiduals}} and a phenotype paths vector for a categorical phenotype made with \code{\link{char2PathsCategorical}}}
\usage{
correlateWithCategoricalPhenotype(
  RERmat,
  charP,
  min.sp = 10,
  min.pos = 2,
  method = "kw",
  winsorizeRER = NULL,
  winsorizetrait = NULL,
  sort = F
)
}
\arguments{
\item{RERmat}{RER matrix returned by \code{\link{getAllResiduals}}}

\item{charP}{phenotype vector returned by \code{\link{tree2Paths}} or \code{\link{char2Paths}}}

\item{min.sp}{Minimum number of species that must be present for a gene}

\item{min.pos}{Minimum number of species that must be present in a category}

\item{method}{Method used to compute correlations. Use "kw" to use Kruskil Wallis. Use "aov" to use ANOVA. It must be one of the strings "kw" or "aov".}

\item{winsorizeRER}{Winsorize RER values before computing Pearson correlation. winsorizeRER=3 will set the 3 most extreme values at each end of each RER row to the the value closest to 0.}

\item{winsorizetrait}{Winsorize trait values before computing Pearson correlation. winsorizetrait=3 will set the 3 most extreme values of the trait values to the value closest to 0.}

\item{sort}{whether to sort by p-value and sign of rho}
}
\value{
A list containing a list object with correlation values, p-values, and the number of data points used for each tree and a list of list objects for each pairwise test with correlation values and p-values.
}
\description{
Computes the association statistics between RER from \code{\link{getAllResiduals}} and a phenotype paths vector for a categorical phenotype made with \code{\link{char2PathsCategorical}}
}
