% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{correlateWithContinuousPhenotype}
\alias{correlateWithContinuousPhenotype}
\title{Computes the association statistics between RER from \code{\link{getAllResiduals}} and a phenotype paths vector for a continuous phenotype made with \code{\link{char2Paths}}}
\usage{
correlateWithContinuousPhenotype(RERmat, charP, min.sp = 10, min.pos = 2,
  winsorize = 3)
}
\arguments{
\item{RERmat}{RER matrix returned by \code{\link{getAllResiduals}}}

\item{charP}{phenotype vector returned by \code{\link{tree2Paths}} or \code{\link{char2Paths}}}

\item{min.sp}{Minimum number of species that must be present for a gene}

\item{winsorize}{Winsorize values before computing Pearson correlation. Winsorize=3, will set the 3 most extreme values at each end to the the value closest to 0.}

\item{method}{Method used to compute correlations. Accepts the same arguments as \code{\link{cor}}. Set to "auto" to select automatically based on the number of unique values in charP. This will also auto set the winsorization for Pearson correlation. Set winsorize=some number to override}
}
\description{
Computes the association statistics between RER from \code{\link{getAllResiduals}} and a phenotype paths vector for a continuous phenotype made with \code{\link{char2Paths}}
}
