% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{edgeVars}
\alias{edgeVars}
\title{maps a vector of traits onto a reference tree}
\usage{
edgeVars(mastertree, tip.vals, metric = "diff", se.filter = -1,
  return.var = F)
}
\value{
A phylo tree with branch values computed from the input tip.values
}
\description{
maps a vector of traits onto a reference tree
}
