% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{foreground2Tree}
\alias{foreground2Tree}
\title{Creates a binary trait tree from a set of foreground species.}
\usage{
foreground2Tree(foreground, treesObj, collapse2anc = T, plotTree = T,
  wholeClade = F, clade = c("ancestral", "terminal", "all",
  "weighted"), useSpecies = NULL)
}
\arguments{
\item{treesObj}{A treesObj created by \code{\link{readTrees}}}

\item{collapse2anc}{Put all the weight on the ancestral branch when the trait appears on a whole clade
(redundant to "clade", kept for backwards compatibility)}

\item{plotTree}{Plot a tree representation of the result}

\item{wholeClade}{Whether to implement the weighted edge option across
all members of a foreground clade (redundant to "clade", kept for backwards compatibility)}

\item{clade}{A character string indicating which branches within the clade
containing the foreground species should be set to foreground. Must be one
of the strings "ancestral", "terminal", "all", or "weighted".}

\item{useSpecies}{Give only a subset of the species to use for ancestral state reconstruction
(e.g., only those species for which the trait can be reliably determined).}

\item{foreground.}{A character vector containing the foreground species}
}
\value{
A tree with edge.lengths representing phenotypic states
}
\description{
Creates a binary trait tree from a set of foreground species.
}
