% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{getAllCor}
\alias{getAllCor}
\title{Computes the association statistics between RER from \code{\link{getAllResiduals}} and a phenotype paths vector made with \code{\link{tree2Paths}} or \code{\link{char2Paths}}}
\usage{
getAllCor(
  RERmat,
  charP,
  method = "auto",
  min.sp = 10,
  min.pos = 2,
  winsorizeRER = NULL,
  winsorizetrait = NULL,
  weighted = F,
  bootstrap = F,
  bootn = 1000,
  sort = F
)
}
\arguments{
\item{RERmat}{RER matrix returned by \code{\link{getAllResiduals}}}

\item{charP}{phenotype vector returned by \code{\link{tree2Paths}} or \code{\link{char2Paths}}}

\item{method}{Method used to compute correlations. Accepts the same arguments as \code{\link{cor}}.}

\item{min.sp}{Minimum number of species that must be present for a gene}

\item{min.pos}{Minimum number of species that must be present in the foreground (non-zero phenotype values)}

\item{winsorizeRER}{Winsorize RER values before computing Pearson correlation. winsorizeRER=3 will set the 3 most extreme RER values at each end of each row to the value closest to 0.}

\item{winsorizetrait}{Winsorize trait values before computing Pearson correlation. winsorizetrait=3 will set the 3 most extreme trait values at each end to the value closest to 0.}

\item{weighted}{perform weighted correlation. This option needs to be set if the clade weights computed in \code{\link{foreground2Tree}(wholeClade=T)} are to be used. This setting will treat the clade a single observation for the purpose of p-value estimation.}

\item{bootstrap}{toggle bootstrapping (for weighted pearson correlation)}

\item{bootn}{number of runs to use when bootstrapping. Will be ignored if bootstrap is false.}

\item{sort}{If set to TRUE, sorts rows by log10(P-value) * -1 \* sign(Rho)}
}
\value{
A list object with correlation values, p-values, and the number of data points used for each tree
}
\description{
Computes the association statistics between RER from \code{\link{getAllResiduals}} and a phenotype paths vector made with \code{\link{tree2Paths}} or \code{\link{char2Paths}}
}
\note{
winsorize is in terms of number of observations at each end, NOT quantiles
}
