% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{getAllCor}
\alias{getAllCor}
\title{Computes the association statistics between RER from \code{\link{getAllResiduals}} and a phenotype paths vector made with \code{\link{tree2Paths}} or \code{\link{char2Paths}}}
\usage{
getAllCor(RERmat, charP, method = "auto", min.sp = 10, min.pos = 2,
  winsorize = NULL, weights = NULL)
}
\arguments{
\item{RERmat}{RER matrix returned by \code{\link{getAllResiduals}}}

\item{charP}{phenotype vector returned by \code{\link{tree2Paths}} or \code{\link{char2Paths}}}

\item{method}{Method used to compute correlations. Accepts the same arguments as \code{\link{cor}}. Set to "auto" to select automatically based on the number of unique values in charP. This will also auto set the winsorization for Pearson correlation. Set winsorize=some number to override}

\item{min.sp}{Minimum number of species that must be present for a gene}

\item{min.pos}{Minimum number of species that must be present in the foreground (non-zero phenotype values)}
}
\value{
A list object with correlation values, p-values, and the number of data points used for each tree
}
\description{
Computes the association statistics between RER from \code{\link{getAllResiduals}} and a phenotype paths vector made with \code{\link{tree2Paths}} or \code{\link{char2Paths}}
}
\note{
winsorize is in terms of number of observations at each end, NOT quantiles
}
