% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{getAllResiduals}
\alias{getAllResiduals}
\title{A wrapper function that runs \code{\link{transformPaths}}, \code{\link{coreGetResiduals}}, and \code{\link{getRMat}}.}
\usage{
getAllResiduals(
  treesObj,
  transform = "sqrt",
  impute = T,
  nvMod = NULL,
  n.pcs = 0,
  cutoff = NULL,
  useSpecies = NULL,
  min.sp = 10,
  min.valid = 20,
  doOnly = NULL,
  maxT = NULL,
  block.do = F,
  weights = NULL,
  use.weights = T,
  interaction = F,
  all = F,
  use.rows = NULL,
  norm = "scale"
)
}
\arguments{
\item{treesObj}{tree object to operate on, product of \code{\link{readTrees}}}

\item{transform}{What transformation to apply. "sqrt" by default.}

\item{impute}{Whether to impute missing data}

\item{nvMod}{A normalization model with nrow(treesObj$paths) rows and any number of columns (optional).}

\item{n.pcs}{Number of principal components to normalize by (default: 0, mean normalization).}

\item{cutoff}{A cutoff for branches to be ignored (optional, lowest 5% by default).}

\item{useSpecies}{Species subset to use (optional).}

\item{min.sp}{Minimum species in a tree.  (default: 10).}

\item{min.valid}{Minimum number of non NA values (after filtering) that must be present for regression to be computed (default: 20).}

\item{doOnly}{Only do specific trees (optional).}

\item{maxT}{Maximum number of trees to do (optional, mostly useful for debugging).}

\item{block.do}{Process rows in blocks (default: FALSE).}

\item{weights}{Weights for calculations (optional, if treesObj has weights those are used).}

\item{use.weights}{Can be used to turn on/off weighted regression.}

\item{interaction}{Use PC interactions when building the normalization model (default: FALSE).}

\item{all}{Logical, indicating whether to return all of the residuals instead of the independent set(default is FALSE).}

\item{use.rows}{A vector of indices of rows/features to return (default is NULL, which means all rows).}

\item{norm}{A character string specifying the normalization method. Options include "scale," "zscore," or "quantile" (default is "scale").}
}
\value{
An residual matrix equivalent to that produced by the original \code{\link{getAllResiduals}}
assuming all=FALSE
}
\description{
A wrapper function that runs \code{\link{transformPaths}}, \code{\link{coreGetResiduals}}, and \code{\link{getRMat}}.
}
