% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{getAllResiduals}
\alias{getAllResiduals}
\title{main RER computation function}
\usage{
getAllResiduals(treesObj, cutoff = 4e-06 * 3, transform = "none",
  weighted = F, useSpecies = NULL, min.sp = 10, scale = F,
  doOnly = NULL, maxT = NULL)
}
\arguments{
\item{treesObj}{A treesObj created by \code{\link{readTrees}}}

\item{transform}{The transformation to apply to the trees branch values before computing relative rates. Available options are sqrt and log, log is recommended.}

\item{weighted}{Use weighted regression to compute relative rates, meant to correct for the non-constant mean-variance relationship in evolutionary rate data}

\item{useSpecies}{Give only a subset of the species to use for RER calculation. Some times excluding unusually long branches can provide more stable results}

\item{min.sp}{The minimum number of species needed to compute RER}

\item{scale}{Scale relative rates internally for each species subset. Increases computation time with little apparent benefit. Better to scale the final matrix.}

\item{doOnly}{The index of a specific tree in the treesObj to calculate RER for. Useful if a single result is needed quickly.}

\item{maxT}{The maximum number of trees to compute results for. Since this function takes some time this is useful for debugging.}

\item{a}{cutoff value for branch lengths bellow which the branch lengths will be discarded, very data dependent but should roughly correspond to 0 or 1 sequence change on that branch}
}
\value{
A numer of trees by number of paths matrix of relative evolutionary rates. Only an independent set of paths has non-NA values for each tree.
}
\description{
main RER computation function
}
