% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{getPermsBinary}
\alias{getPermsBinary}
\title{Calculates permuted correlation and enrichment statistics for binary phenotype}
\usage{
getPermsBinary(
  numperms,
  fg_vec,
  sisters_list,
  root_sp,
  RERmat,
  trees,
  mastertree,
  permmode = "cc",
  method = "k",
  trees_list = NULL,
  calculateenrich = F,
  annotlist = NULL
)
}
\arguments{
\item{numperms}{An integer number of permulations}

\item{fg_vec}{A vector containing the foreground species}

\item{sisters_list}{A list containing pairs of "sister species" in the foreground set (put NULL if empty)}

\item{root_sp}{The species to root the tree on}

\item{RERmat}{An RER matrix calculated using \code{\link{getAllResiduals}}.}

\item{trees}{treesObj from \code{\link{readTrees}}}

\item{mastertree}{A rooted, fully dichotomous tree derived from the treesObj master tree from \code{\link{readTrees}}.  Must not contain species not in traitvec}

\item{permmode}{Mode of binary permulation ("cc" for Complete Cases (default), "ssm" for Species Subset Match)}

\item{method}{statistical method to use for correlations (set to "k" (default) for Kendall Tau test)}

\item{calculateenrich}{A boolean variable indicating if null permulation p-values for enrichment statistics}

\item{annotlist}{Pathway annotations}
}
\value{
A list object with enrichment statistics, correlation p-val, rho, and correlation effect size
}
\description{
Calculates permuted correlation and enrichment statistics for binary phenotype
}
