% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{namePathsWSpecies}
\alias{namePathsWSpecies}
\title{Provides names for paths/RERs representing terminal branches for plotting
Originally an internal function but necessary for the vignette/walk-through}
\usage{
namePathsWSpecies(masterTree)
}
\arguments{
\item{masterTree}{The master tree used for analysis}
}
\value{
Names corresponding to the paths/RERs for terminal branches
}
\description{
Provides names for paths/RERs representing terminal branches for plotting
Originally an internal function but necessary for the vignette/walk-through
}
