% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{permpvalcor}
\alias{permpvalcor}
\title{Calculates correlation permutation pvals from output of \code{\link{getPermsContinuous}}}
\usage{
permpvalcor(realcor, permvals)
}
\arguments{
\item{realcor}{Real enrichment statistics from \code{\link{fastwilcoxGMTall}}}

\item{permvals}{output from \code{\link{getPermsContinuous}}}
}
\value{
A vector with permulation p-values
}
\description{
Calculates correlation permutation pvals from output of \code{\link{getPermsContinuous}}
}
