% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermulationFuncs.R
\name{permpvalenrich}
\alias{permpvalenrich}
\title{Calculates permutation pvals from output of \code{\link{getPerms}}}
\usage{
permpvalenrich(realenrich, permenrich)
}
\arguments{
\item{realenrich}{Real enrichment statistics from \code{\link{fastwilcoxGMTall}}}

\item{permenrich}{signedenrichmentstats from \code{\link{getPerms}}}
}
\value{
A list object with permulation p-values
}
\description{
Calculates permutation pvals from output of \code{\link{getPerms}}
}
