% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{readTrees}
\alias{readTrees}
\title{reads trees from a 2 column , tab seperated, file
The first columns is the gene name and the second column is the corresponding tree in parenthetic format known as the Newick or New Hampshire format}
\usage{
<<<<<<< HEAD
readTrees(file, max.read = NA, masterTree = NULL, minTreesAll = 20)
=======
readTrees(file, max.read = NA, reorient = F, outgroup = NULL)
>>>>>>> 3fcc6028e858e410a9bcaff4a8459dce0027e84e
}
\arguments{
\item{file}{The path to the tree file}

<<<<<<< HEAD
\item{max.read}{This function takes a while for a whole genome, so max.read is useful for testing}

\item{masterTree}{(optional) User can specify a master tree; only the topology will be used, and branch lengths will be inferred from gene trees.}

\item{minTreesAll}{The minimum number of trees with all species present in order to estimate
master tree edge lengths (default 20).}

\item{masterTree}{(optional) User can specify a master tree. Recommended only when
the number of available gene trees with all species is small.}
=======
\item{max.read}{this function takes a while for  a whole genome so max.read is useful for testing}

\item{reorient}{Whether to re-orient the master tree so that it is pseudo-rooted
on an outgroup. If unspecified in `outgroup`, this will use `midpoint.root`.}

\item{outgroup}{A user-specified outgroup for re-rooting with `root`
(only used if reorient=TRUE)}
>>>>>>> 3fcc6028e858e410a9bcaff4a8459dce0027e84e
}
\value{
A trees object of class "treeObj"

A trees object of class "treeObj"
}
\description{
reads trees from a 2 column , tab seperated, file
The first columns is the gene name and the second column is the corresponding tree in parenthetic format known as the Newick or New Hampshire format
}
