% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{readTrees}
\alias{readTrees}
\title{reads trees from a 2 column , tab seperated, file
The first columns is the gene name and the second column is the corresponding tree in parenthetic format known as the Newick or New Hampshire format}
\usage{
readTrees(file, max.read = NA)
}
\arguments{
\item{file}{The path to the tree file}

\item{max.read}{this function takes a while for  a whole genome so max.read is useful for testing}
}
\value{
A trees object of class "treeObj"
}
\description{
reads trees from a 2 column , tab seperated, file
The first columns is the gene name and the second column is the corresponding tree in parenthetic format known as the Newick or New Hampshire format
}
