% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{transformPaths}
\alias{transformPaths}
\title{Apply a transformation to a trees object}
\usage{
transformPaths(treesObj, transform = "sqrt", impute = T)
}
\arguments{
\item{transform}{What transformation to apply. "sqrt" by default.}

\item{impute}{Whether to impute missing data}
}
\value{
A treesObj with transformed paths
}
\description{
Apply a transformation to a trees object
}
