% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RERfuncs.R
\name{transformPaths}
\alias{transformPaths}
\title{Transform Paths}
\usage{
transformPaths(treesObj, transform = "log", impute = T)
}
\arguments{
\item{treesObj}{A tree object}

\item{transform}{A character string specifying the transformation to apply.
Options include "log", "sqrt", "exp", etc.}

\item{impute}{Logical, indicating whether to impute missing values (default is TRUE).}
}
\value{
A modified tree object with transformed paths.
}
\description{
This function applies a transformation to the paths in a tree object optionally imputing missing values
}
